requirements.txt
langchain==0.0.166
llama-cpp-python
urllib3==1.26.6
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc-binary==1.11
torch==2.0.0 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
streamlit==1.25.0
bitsandbytes==0.38.1
qdrant-client==1.3.1